check box to black
