+++ app/views/devise/sessions/new.html.erb
@@ -1,5 +1,5 @@
 
- <div class="container m-5"> <h1>Heureux de vous revoir</h1> </div>
+ <div class="container mt-3 mb-3"> <h1>Heureux de vous revoir</h1> </div>
 <div class="container sign-in-container">
 
   <h2>Connexion</h2>
