+++ app/views/layouts/application.html.erb
@@ -13,7 +13,8 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,600;1,600&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=Montserrat:ital@0;1&display=swap" rel="stylesheet">
-  
+    
+
    <script src="https://kit.fontawesome.com/45e38e596f.js" crossorigin="anonymous"></script>
 
   </head>
