+++ app/views/lists/show.html.erb
@@ -1,8 +1,9 @@
 <div class="container">
-  <div class="list-card-product">
+  <div class="list-card-product" style="background-image: url(<%= image_path(image_category(@list.category)) %>)">
     <div class="row">
 
-      <h2 class="text-center"><ul><%= @list.name %></ul></h2>
+      <h2 class="text-center">
+        <%= @list.name %></h2>
           <% @list.order_by_done_tasks.each do |task| %>
             <%= simple_form_for task, data: { controller: "submit-task" } do |f| %>
               <%= f.input :done, as: :boolean, label: task.label, input_html: {data: {action: "change->submit-task#submitOnChange"}} %>
