+++ app/views/shared/_navbar.html.erb
@@ -2,7 +2,7 @@
 
   <div class="container-fluid">
     <%= link_to root_path, class: "navbar-brand p-0" do %>
-    <%=image_tag "logo-t3.png", alt: "logo", width: "80" %>
+    <%=image_tag "logo-t4.png", alt: "logo", width: "50" %>
     <% end %>
 
     <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
